feat(gui): add button to open user data folder (closes #97)
